requirements.txt
python-dotenv
django
djangorestframework
djangorestframework-simplejwt
django-redis
mysqlclient
cryptography
django-cors-headers